localcert/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.7.1
Django==5.0.7
django-allauth==0.63.3
django-csp==3.8
dnspython==2.6.1
gunicorn==22.0.0
idna==3.7
oauthlib==3.2.2
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.0
urllib3==2.2.2
